error test JPL con tupla

+++ src/Cheb3D.cpp
@@ -1,6 +1,6 @@
 #include "../include/Cheb3D.hpp"
 
-Matrix& Cheb3D(const int t,const int N,const int Ta,const int Tb,Matrix &Cx,Matrix &Cy,Matrix &Cz){
+Matrix& Cheb3D(const double t,const int N,const double Ta,const double Tb,Matrix &Cx,Matrix &Cy,Matrix &Cz){
 
 //Check validity
 if ( (t<Ta) || (Tb<t) ){
